cleaned up and reorganized the navbar code, also tested the signup route was working.
