+++ client/src/App.js
@@ -2,7 +2,8 @@ import React from "react";
 import { BrowserRouter as Router, Route, Switch } from "react-router-dom";
 import Login from "./components/Login";
 import Home from './pages/home';
-import MyNav from "./pages/navbar"
+import MyNav from "./components/Nav/navbar"
+import SignUp from "./pages/signup"
 import 'bootstrap/dist/css/bootstrap.min.css';
 
 
@@ -12,6 +13,7 @@ function App() {
       <MyNav/>
       <div>
       <Route exact path="/" component={Home}/>
+      <Route exact path="/signup" component={SignUp}/>
       </div>
     </Router>
