+++ client/src/pages/signup.js
@@ -1,0 +1,12 @@
+import React from 'react';
+
+ 
+function SignUp () {
+    return (
+        <div>
+           initial sign up page, testing if navbar route works.
+        </div>
+    )
+}
+
+export default SignUp;
